+++ src/main/java/com/expense_tracker/user/UserInfoService.java
@@ -32,10 +32,10 @@ public class UserInfoService implements UserDetailsService {
 
 	public String addUser(UserInfo userInfo) {
 		// verify if username or email already exist
-		if (repository.existsByEmail(userInfo.getEmail())) {
+		if (Boolean.TRUE.equals(repository.existsByEmail(userInfo.getEmail()))) {
 			throw new DuplicateKeyException("Email already exists :" + userInfo.getEmail());
 		}
-		if (repository.existsByUsername((userInfo.getUsername()))) {
+		if (Boolean.TRUE.equals(repository.existsByUsername((userInfo.getUsername())))) {
 			throw new DuplicateKeyException("Username already exists :" + userInfo.getUsername());
 		}
 		// Encode password before saving the user
